Добавление моделей данных, правки по моделям данных.

+++ src/main/java/ru/job4j/cars/model/PriceHistory.java
@@ -15,10 +15,12 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PriceHistory {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @EqualsAndHashCode.Include
-    private int id;
+    private Long id;
+
     private long before;
     private long after;
     private LocalDateTime created;
